refactor(v2.10): slim spec.rs to idl.rs; drop SPEC.md generation
Group B from PRD-v2.10.md — release-shaped cleanup that reinforces
the .qedspec-as-front-door framing the auditor leans on. The SPEC.md
generators that previously lived in `crates/qedgen/src/spec.rs` are
removed entirely; what's kept is the IDL parsing and pattern inference
layer that other modules (idl2spec, interface_gen) actually consume.

- B1 — `spec.rs` (847 LOC) slim-and-rename to `idl.rs` (~210 LOC).
  Keepers: `Idl` and the typed IDL substructure
  (`IdlMetadata`/`IdlInstruction`/`IdlAccount`/`IdlPda`/`IdlSeed`/
  `IdlArg`/`IdlTypeDef`/`IdlTypeBody`/`IdlField`/`IdlError`),
  `parse_idl`, `analyze_instruction`, `InstructionAnalysis`,
  `type_label`, `snake_to_title`. Deleted: `generate_spec`,
  `generate_spec_from_qedspec`, `build_spec` (~660 LOC of Markdown
  narrative generation).

- B2 — `qedgen spec` clap surface trimmed. Dropped flags:
  `--from-spec`, `--proofs`, `--format`. The subcommand is now
  IDL → `.qedspec` exclusively (calls `idl2spec::generate_qedspec`).
  Breaking change for any caller depending on the SPEC.md default
  output, which is why this lands at a version boundary per
  `feedback_minor_release_completeness.md`.

- B3 — Deleted 5 stale committed SPEC.md fixtures (escrow, percolator,
  dropset, slippage, transfer). Each had drifted from its `.qedspec`
  by weeks; nothing reads them.

- B4 — Docs sweep:
  - README.md: drop `--format qedspec` flag from invocation examples
  - CLAUDE.md: change "Generate a draft SPEC.md" comment to "Scaffold
    a .qedspec"; update Architecture section to describe `idl.rs`
    instead of `spec.rs`
  - references/cli.md: rewrite the `spec` section to reflect
    IDL → `.qedspec` only; explicit historical note that SPEC.md
    generators were dropped in v2.10
  - claude.md: identical to CLAUDE.md (case-insensitive filesystem
    on macOS resolves both to the same file)

Module rename: `crate::spec::*` → `crate::idl::*` across
`interface_gen.rs` and `idl2spec.rs`. `mod spec;` → `mod idl;` in
`main.rs`.

Cleanups discovered during the rename:
- `InstructionAnalysis` fields previously consumed only by the
  SPEC.md narrative (`display_name`, `has_one_relations`, `args`,
  `has_numeric_args`) tagged `#[allow(dead_code)]` — kept on the
  struct as stable analysis surface for future scaffolders.
- `ParsedHandler::has_when` similarly tagged (only consumer was
  `generate_spec_from_qedspec`).

Net: 1568 lines deleted, 46 added across 13 files. Tests: 434
passing. Probe regression check on showcase specs: clean.
qedgen check on multisig + lending: only preexisting P2/P3 lints
(unrelated to this commit).

Closes Group B per PRD-v2.10.md. Group C (cross-harness clarity)
already shipped via the auditor SKILL.md install. Remaining v2.10
work: A5 (pre-launch eval pass on the 11-program eval set) and
verifying all shipped examples are probe-clean / bug-free per the
v2.10 hygiene framing.
